tucana: Pretend to be launched with dynamic partitions
A dist build fails on a-only devices with retrofit dynamic partitions due to [1].
The check does not take a-only devices into account, even though they are capable of retrofit aswell.
While the PRODUCT_RETROFIT_DYNAMIC_PARTITIONS flag is used to detect and warn about misconfigurations, it does not seem to be needed for any retrofit functionality.
Hence, to avoid changes in releasetools, drop the flag.

[1]: https://github.com/LineageOS/android_build/blob/055b7692fb8b8830991774ed0d68a94abd76c335/tools/releasetools/check_partition_sizes.py#L185

Change-Id: Ib0dfe4ee3da4c583d924d85a79d5ec37a2f9ba87